dom/alarm/test/test_bug1090896.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test time alert is fired for Bug 1090896</title>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test">
  <script type="application/javascript">

  "use strict";

  function registerPage() {
    var gScript = SpecialPowers.loadChromeScript(SimpleTest.getTestFileURL('system_message_chrome_script.js'));
    gScript.addMessageListener("page-registered", function pageRegisteredHandler() {
      gScript.removeMessageListener("page-registered", pageRegisteredHandler);
      gScript.destroy();
      testFireTimeAlertWithNoData();
    });

    gScript.sendAsyncMessage("trigger-register-page",
                             { type: "alarm",
                               manifestURL: window.location.origin + "/manifest.webapp",
                               pageURL: window.location.href });
  }

  function testFireTimeAlertWithNoData() {
    var secondsLater = new Date();
    secondsLater.setSeconds(secondsLater.getSeconds() + 1);

    var domRequest;
    try {
      // Set system message handler.
      navigator.mozSetMessageHandler('alarm', function(message){
        ok(true, "Time alert has been fired.");
        SimpleTest.finish();
      });

      domRequest = navigator.mozAlarms.add(secondsLater, "honorTimezone");
    } catch (e) {
      ok(false, "Unexpected exception trying to set time alert. " + e);

      return SimpleTest.finish();
    }
    domRequest.onsuccess = function(e) {
      ok(true, "Set time alert. Waiting to be fired.");
    };
    domRequest.onerror = function(e) {
      ok(false, "Unable to set time alert.");

      SimpleTest.finish();
    };
  }

  function startTests() {
    SpecialPowers.pushPrefEnv({
      "set": [["dom.mozAlarms.enabled", true]]
    }, function() {
      var isAllowedToTest = true;

      if (navigator.appVersion.indexOf("Android") !== -1) {
        ok(true, "mozAlarms is not allowed on Android for now. " +
                 "TODO Bug 863557.");
        isAllowedToTest = false;
      } else if (SpecialPowers.wrap(document).nodePrincipal.appStatus ==
                 SpecialPowers.Ci.nsIPrincipal.APP_STATUS_NOT_INSTALLED) {
        ok(true, "mozAlarms is not allowed for non-installed apps. " +
                 "TODO Bug 876981.");
        isAllowedToTest = false;
      }

      if (isAllowedToTest) {
        ok(true, "Start to test...");
        registerPage();
      } else {
        // A sanity check to make sure we must run tests on Firefox OS (B2G).
        if (navigator.userAgent.indexOf("Mobile") != -1 &&
            navigator.appVersion.indexOf("Android") == -1) {
          ok(false, "Should run the test on Firefox OS (B2G)!");
        }

        SimpleTest.finish();
      }
    });
  }

  SimpleTest.waitForExplicitFinish();
  SpecialPowers.pushPermissions([{'type': 'alarms', 'allow': true, 'context': document}], startTests);

  </script>
</pre>
</body>
</html>
